source/requirements.txt
absl-py==2.1.0
accelerate==1.6.0
adapters==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.0
bleach==6.1.0
blobfile==3.0.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.3.2
debugpy==1.8.6
decorator==5.1.1
deepspeed==0.12.6
deformable-attention==0.0.19
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.5
fastapi==0.115.0
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.17.0
fire==0.6.0
flash-attn @ file:///home/ubuntu/wxl/LLaVA_error/flash_attn-2.6.3%2Bcu118torch2.1cxx11abiFALSE-cp310-cp310-linux_x86_64.whl#sha256=d16f6008f2b7727718653bfd087a20f3e18e2a4c8e91b9f38da2439f3b0bf529
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.66.1
h5py==3.12.1
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.31.2
idna==3.10
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.14.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
jiter==0.7.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
latex2mathml==3.77.0
llama_models==0.0.19
llama_toolchain==0.0.17
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpi4py @ file:///home/conda/feedstock_root/build_artifacts/mpi4py_1713171790760/work
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
MultiScaleDeformableAttention==1.0
narwhals==1.8.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
onedrivedownloader==1.1.3
openai @ file:///home/ubuntu/jpk/llava/LLaVA_RoDE/openai-1.54.4-py3-none-any.whl#sha256=0d95cef99346bf9b6d7fbf57faf61a673924c3e34fa8af84c9ffe04660673a7e
orjson==3.10.7
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
prometheus_client==0.21.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==6.31.1
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydub==0.25.1
Pygments==2.18.0
pynvml==11.5.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rouge==1.0.1
rouge-score==0.1.2
rpds-py==0.20.0
rsa==4.9
ruff==0.6.7
sacrebleu==2.4.3
safetensors==0.4.3
scikit-learn==1.2.2
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.1.99
sentry-sdk==2.14.0
shap==0.46.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.38.6
svgwrite==1.4.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==0.6.13
tinycss2==1.3.0
tokenizers==0.15.2
tomli==2.0.2
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
triton==2.1.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.30.6
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
widgetsnbextension==4.0.13
xxhash==3.5.0
yarl==1.11.1
yaspin==3.0.2
